[REF] trainingapp: Refactor js and add filters
This commit removes simple data container classes Training and Dashboard with plain data objects stored in session.data.
The rationale is simply that since they have no special behaviors, there is no need to have a class. All manipulations
are performed server-side in advance by json controllers, providing the data in the exact form needed for the respective
wigets.

Additionally, the User class has been renamed to TrainingAppSession, and user_id is an attribute of the Session, to
reinforce the semantic distinction between the app doing work with user as an argument vs the old way of the user
doing the work. This will be useful for apps using user data.

Further, this commit adds logic to filter the accordian based on the users clicks